Windows version of ofxPuppet by Bryce Summers, April 2014. Supported by the Frank-Ratchye STUDIO for Creative Inquiry at Carnegie Mellon University
I have created a Visual Studio Project that uses the same source as the
xCode example project almost verbatim, except for two lines that
indicate how many microseconds the Singular Value Decomposition
Operation takes to perform.

The code builds on my machine when it is in the normal Openframeworks
app/myApps folder.